% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-tokens.R
\name{fix_quote}
\alias{fix_quote}
\title{Replace double quotes with single quotes}
\usage{
fix_quote(pd_flat)
}
\arguments{
\item{pd_flat}{A flat parse table.}
}
\description{
Replace double quotes with single quotes
}
\examples{
style_text("'here is a string'", style = quote_style)
}
\keyword{internal}
